makeBet tests are working
