--- test/logic.js
@@ -1,105 +1,0 @@
-'use strict';
-/**
- * Write the unit tests for your transction processor functions here
- */
-
-const AdminConnection = require('composer-admin').AdminConnection;
-const BusinessNetworkConnection = require('composer-client').BusinessNetworkConnection;
-const BusinessNetworkDefinition = require('composer-common').BusinessNetworkDefinition;
-const IdCard = require('composer-common').IdCard;
-const MemoryCardStore = require('composer-common').MemoryCardStore;
-
-const path = require('path');
-
-require('chai').should();
-
-const namespace = 'com.omni.biznet';
-
-describe('#' + namespace, () => {
-    // In-memory card store for testing so cards are not persisted to the file system
-    const cardStore = new MemoryCardStore();
-    let adminConnection;
-    let businessNetworkConnection;
-
-    before(() => {
-        // Embedded connection used for local testing
-        const connectionProfile = {
-            name: 'embedded',
-            type: 'embedded'
-        };
-        // Embedded connection does not need real credentials
-        const credentials = {
-            certificate: 'FAKE CERTIFICATE',
-            privateKey: 'FAKE PRIVATE KEY'
-        };
-
-        // PeerAdmin identity used with the admin connection to deploy business networks
-        const deployerMetadata = {
-            version: 1,
-            userName: 'PeerAdmin',
-            roles: [ 'PeerAdmin', 'ChannelAdmin' ]
-        };
-        const deployerCard = new IdCard(deployerMetadata, connectionProfile);
-        deployerCard.setCredentials(credentials);
-
-        const deployerCardName = 'PeerAdmin';
-        adminConnection = new AdminConnection({ cardStore: cardStore });
-
-        return adminConnection.importCard(deployerCardName, deployerCard).then(() => {
-            return adminConnection.connect(deployerCardName);
-        });
-    });
-
-    beforeEach(() => {
-        businessNetworkConnection = new BusinessNetworkConnection({ cardStore: cardStore });
-
-        const adminUserName = 'admin';
-        let adminCardName;
-        let businessNetworkDefinition;
-
-        return BusinessNetworkDefinition.fromDirectory(path.resolve(__dirname, '..')).then(definition => {
-            businessNetworkDefinition = definition;
-            // Install the Composer runtime for the new business network
-            return adminConnection.install(businessNetworkDefinition.getName());
-        }).then(() => {
-            // Start the business network and configure an network admin identity
-            const startOptions = {
-                networkAdmins: [
-                    {
-                        userName: adminUserName,
-                        enrollmentSecret: 'adminpw'
-                    }
-                ]
-            };
-            return adminConnection.start(businessNetworkDefinition, startOptions);
-        }).then(adminCards => {
-            // Import the network admin identity for us to use
-            adminCardName = `${adminUserName}@${businessNetworkDefinition.getName()}`;
-            return adminConnection.importCard(adminCardName, adminCards.get(adminUserName));
-        }).then(() => {
-            // Connect to the business network using the network admin identity
-            return businessNetworkConnection.connect(adminCardName);
-        });
-    });
-
-    describe('RegisterUser()', () => {
-        it('should register a new user with 1000 starting balance', () => {
-            const factory = businessNetworkConnection.getBusinessNetwork().getFactory();
-
-            const email = 'user1';
-            const registerUserTx = factory.newTransaction(namespace, 'RegisterUser');
-            registerUserTx.email = email;
-            registerUserTx.charity = 'Undefined';
-
-            return Promise.all([
-              businessNetworkConnection.getParticipantRegistry(`${namespace}.User`),
-              businessNetworkConnection.submitTransaction(registerUserTx)
-            ]).then(([ userRegistry, tx ]) => {
-              return userRegistry.get(email);
-            }).then(user => {
-              user.balance.should.equal(1000);
-            });
-        });
-    });
-
-});
